flask-server/requirements.txt
flask~=2.0.3
pytorch~=1.10.2
tokenizers~=0.11.4
transformers~=4.18.0